feat: Add pros-cons-discusser-steam use case and update use-case index
This commit adds the pros-cons-discusser-steam use case to the use-case index file. It also includes the necessary code for making a POST request to the pros-cons-discusser-stream endpoint. This use case allows for discussing pros and cons using a streaming response.

The use-case index file is updated to export the new use case along with the existing ones. This change enhances the functionality of the application by providing a more efficient way to handle pros and cons discussions.