feat: Add audio-to-text use case and update use-case index
This commit adds the audio-to-text use case to the use-case index file. It also includes the necessary code for converting audio to text using the OpenAI API. By adding this functionality, users can now convert audio files to text using the application. This enhances the versatility of the application by providing a convenient way to transcribe audio content.