feat: Update TextMessageBoxFile to handle audio file in onSendMessage
This commit updates the TextMessageBoxFile component to handle an additional parameter, audioFile, in the onSendMessage function. Previously, the onSendMessage function only accepted a message parameter, but now it also accepts an audioFile parameter. This change allows the component to send both the message and the selected audio file when the form is submitted.

By adding support for audio files in the onSendMessage function, users can now send audio messages in addition to text messages. This enhances the functionality of the component and provides a more versatile messaging experience.