feat: Add downloadImageAsPng and downloadBase64ImageAsPng helper functions
These helper functions are added to the `helpers` directory. The `downloadImageAsPng` function downloads an image from a given URL and saves it as a PNG file. The `downloadBase64ImageAsPng` function converts a base64-encoded image to a PNG file. These functions enhance the functionality of the application by providing convenient ways to download and convert images.