chore: Update ESLint configuration to disable 'no-unused-vars' rule and 'no-non-null-asserted-optional-chain' rule
The ESLint configuration file (.eslintrc.cjs) has been updated to disable the 'no-unused-vars' rule and the 'no-non-null-asserted-optional-chain' rule. This change allows unused variables to trigger a warning instead of an error, and allows the use of non-null asserted optional chaining. These updates improve the code quality and flexibility of the project.